feat: 优化章节详情的数据处理，后端仅返回前端所需的Episode列表

+++ src-tauri/src/responses/manga_data.rs
@@ -6,14 +6,14 @@ use specta::Type;
 #[allow(clippy::struct_excessive_bools)]
 #[allow(clippy::struct_field_names)]
 pub struct MangaData {
-    pub id: i32,
+    pub id: i64,
     pub title: String,
     #[serde(rename = "comic_type")]
-    pub comic_type: i32,
+    pub comic_type: i64,
     #[serde(rename = "page_default")]
-    pub page_default: i32,
+    pub page_default: i64,
     #[serde(rename = "page_allow")]
-    pub page_allow: i32,
+    pub page_allow: i64,
     #[serde(rename = "horizontal_cover")]
     pub horizontal_cover: String,
     #[serde(rename = "square_cover")]
@@ -24,27 +24,27 @@ pub struct MangaData {
     pub author_name: Vec<String>,
     pub styles: Vec<String>,
     #[serde(rename = "last_ord")]
-    pub last_ord: i32,
+    pub last_ord: i64,
     #[serde(rename = "is_finish")]
-    pub is_finish: i32,
-    pub status: i32,
-    pub fav: i32,
+    pub is_finish: i64,
+    pub status: i64,
+    pub fav: i64,
     #[serde(rename = "read_order")]
-    pub read_order: i32,
+    pub read_order: i64,
     pub evaluate: String,
-    pub total: i32,
+    pub total: i64,
     #[serde(rename = "ep_list")]
     pub ep_list: Vec<EpList>,
     #[serde(rename = "release_time")]
     pub release_time: String,
     #[serde(rename = "is_limit")]
-    pub is_limit: i32,
+    pub is_limit: i64,
     #[serde(rename = "read_epid")]
-    pub read_epid: i32,
+    pub read_epid: i64,
     #[serde(rename = "last_read_time")]
     pub last_read_time: String,
     #[serde(rename = "is_download")]
-    pub is_download: i32,
+    pub is_download: i64,
     #[serde(rename = "read_short_title")]
     pub read_short_title: String,
     pub styles2: Vec<Styles2>,
@@ -53,53 +53,53 @@ pub struct MangaData {
     #[serde(rename = "last_short_title")]
     pub last_short_title: String,
     #[serde(rename = "discount_type")]
-    pub discount_type: i32,
-    pub discount: i32,
+    pub discount_type: i64,
+    pub discount: i64,
     #[serde(rename = "discount_end")]
     pub discount_end: String,
     #[serde(rename = "no_reward")]
     pub no_reward: bool,
     #[serde(rename = "batch_discount_type")]
-    pub batch_discount_type: i32,
+    pub batch_discount_type: i64,
     #[serde(rename = "ep_discount_type")]
-    pub ep_discount_type: i32,
+    pub ep_discount_type: i64,
     #[serde(rename = "has_fav_activity")]
     pub has_fav_activity: bool,
     #[serde(rename = "fav_free_amount")]
-    pub fav_free_amount: i32,
+    pub fav_free_amount: i64,
     #[serde(rename = "allow_wait_free")]
     pub allow_wait_free: bool,
     #[serde(rename = "wait_hour")]
-    pub wait_hour: i32,
+    pub wait_hour: i64,
     #[serde(rename = "wait_free_at")]
     pub wait_free_at: String,
     #[serde(rename = "no_danmaku")]
-    pub no_danmaku: i32,
+    pub no_danmaku: i64,
     #[serde(rename = "auto_pay_status")]
-    pub auto_pay_status: i32,
+    pub auto_pay_status: i64,
     #[serde(rename = "no_month_ticket")]
     pub no_month_ticket: bool,
     pub immersive: bool,
     #[serde(rename = "no_discount")]
     pub no_discount: bool,
     #[serde(rename = "show_type")]
-    pub show_type: i32,
+    pub show_type: i64,
     #[serde(rename = "pay_mode")]
-    pub pay_mode: i32,
+    pub pay_mode: i64,
     #[serde(rename = "classic_lines")]
     pub classic_lines: String,
     #[serde(rename = "pay_for_new")]
-    pub pay_for_new: i32,
+    pub pay_for_new: i64,
     #[serde(rename = "fav_comic_info")]
     pub fav_comic_info: FavComicInfo,
     #[serde(rename = "serial_status")]
-    pub serial_status: i32,
+    pub serial_status: i64,
     #[serde(rename = "album_count")]
-    pub album_count: i32,
+    pub album_count: i64,
     #[serde(rename = "wiki_id")]
-    pub wiki_id: i32,
+    pub wiki_id: i64,
     #[serde(rename = "disable_coupon_amount")]
-    pub disable_coupon_amount: i32,
+    pub disable_coupon_amount: i64,
     #[serde(rename = "japan_comic")]
     pub japan_comic: bool,
     #[serde(rename = "interact_value")]
@@ -108,27 +108,27 @@ pub struct MangaData {
     pub temporary_finish_time: String,
     pub introduction: String,
     #[serde(rename = "comment_status")]
-    pub comment_status: i32,
+    pub comment_status: i64,
     #[serde(rename = "no_screenshot")]
     pub no_screenshot: bool,
     #[serde(rename = "type")]
-    pub type_field: i32,
+    pub type_field: i64,
     #[serde(rename = "no_rank")]
     pub no_rank: bool,
     #[serde(rename = "presale_text")]
     pub presale_text: String,
     #[serde(rename = "presale_discount")]
-    pub presale_discount: i32,
+    pub presale_discount: i64,
     #[serde(rename = "no_leaderboard")]
     pub no_leaderboard: bool,
     #[serde(rename = "auto_pay_info")]
     pub auto_pay_info: AutoPayInfo,
-    pub orientation: i32,
+    pub orientation: i64,
     #[serde(rename = "story_elems")]
     pub story_elems: Vec<StoryElem>,
     pub tags: Vec<Tag>,
     #[serde(rename = "is_star_hall")]
-    pub is_star_hall: i32,
+    pub is_star_hall: i64,
     #[serde(rename = "hall_icon_text")]
     pub hall_icon_text: String,
     #[serde(rename = "rookie_fav_tip")]
@@ -147,16 +147,16 @@ pub struct MangaData {
 #[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize, Type)]
 #[serde(rename_all = "camelCase")]
 pub struct EpList {
-    pub id: i32,
+    pub id: i64,
     pub ord: f64,
-    pub read: i32,
+    pub read: i64,
     #[serde(rename = "pay_mode")]
-    pub pay_mode: i32,
+    pub pay_mode: i64,
     #[serde(rename = "is_locked")]
     pub is_locked: bool,
     #[serde(rename = "pay_gold")]
-    pub pay_gold: i32,
-    pub size: i32,
+    pub pay_gold: i64,
+    pub size: i64,
     #[serde(rename = "short_title")]
     pub short_title: String,
     #[serde(rename = "is_in_free")]
@@ -165,23 +165,23 @@ pub struct EpList {
     pub cover: String,
     #[serde(rename = "pub_time")]
     pub pub_time: String,
-    pub comments: i32,
+    pub comments: i64,
     #[serde(rename = "unlock_expire_at")]
     pub unlock_expire_at: String,
     #[serde(rename = "unlock_type")]
-    pub unlock_type: i32,
+    pub unlock_type: i64,
     #[serde(rename = "allow_wait_free")]
     pub allow_wait_free: bool,
     pub progress: String,
     #[serde(rename = "like_count")]
-    pub like_count: i32,
+    pub like_count: i64,
     #[serde(rename = "chapter_id")]
-    pub chapter_id: i32,
+    pub chapter_id: i64,
     #[serde(rename = "type")]
-    pub type_field: i32,
-    pub extra: i32,
+    pub type_field: i64,
+    pub extra: i64,
     #[serde(rename = "image_count")]
-    pub image_count: i32,
+    pub image_count: i64,
     #[serde(rename = "index_last_modified")]
     pub index_last_modified: String,
     #[serde(rename = "jump_url")]
@@ -191,7 +191,7 @@ pub struct EpList {
 #[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize, Type)]
 #[serde(rename_all = "camelCase")]
 pub struct Styles2 {
-    pub id: i32,
+    pub id: i64,
     pub name: String,
 }
 
@@ -201,9 +201,9 @@ pub struct FavComicInfo {
     #[serde(rename = "has_fav_activity")]
     pub has_fav_activity: bool,
     #[serde(rename = "fav_free_amount")]
-    pub fav_free_amount: i32,
+    pub fav_free_amount: i64,
     #[serde(rename = "fav_coupon_type")]
-    pub fav_coupon_type: i32,
+    pub fav_coupon_type: i64,
 }
 
 #[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize, Type)]
@@ -211,27 +211,27 @@ pub struct FavComicInfo {
 pub struct AutoPayInfo {
     #[serde(rename = "auto_pay_orders")]
     pub auto_pay_orders: Vec<AutoPayOrder>,
-    pub id: i32,
+    pub id: i64,
 }
 
 #[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize, Type)]
 #[serde(rename_all = "camelCase")]
 pub struct AutoPayOrder {
-    pub id: i32,
+    pub id: i64,
     pub title: String,
 }
 
 #[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize, Type)]
 #[serde(rename_all = "camelCase")]
 pub struct StoryElem {
-    pub id: i32,
+    pub id: i64,
     pub name: String,
 }
 
 #[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize, Type)]
 #[serde(rename_all = "camelCase")]
 pub struct Tag {
-    pub id: i32,
+    pub id: i64,
     pub name: String,
 }
 
@@ -240,14 +240,14 @@ pub struct Tag {
 pub struct RookieFavTip {
     #[serde(rename = "is_show")]
     pub is_show: bool,
-    pub used: i32,
-    pub total: i32,
+    pub used: i64,
+    pub total: i64,
 }
 
 #[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize, Type)]
 #[serde(rename_all = "camelCase")]
 pub struct Author {
-    pub id: i32,
+    pub id: i64,
     pub name: String,
     pub cname: String,
 }
@@ -277,9 +277,9 @@ pub struct ReadScore {
 #[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize, Type)]
 #[serde(rename_all = "camelCase")]
 pub struct Increase {
-    pub days: i32,
+    pub days: i64,
     #[serde(rename = "increase_percent")]
-    pub increase_percent: i32,
+    pub increase_percent: i64,
 }
 
 #[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize, Type)]
@@ -297,7 +297,7 @@ pub struct InteractiveValue {
 #[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize, Type)]
 #[serde(rename_all = "camelCase")]
 pub struct Increase2 {
-    pub days: i32,
+    pub days: i64,
     #[serde(rename = "increase_percent")]
-    pub increase_percent: i32,
+    pub increase_percent: i64,
 }
